+++ packages/mcp/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.2.2 (2025-10-07)
+
+### 🧱 Updated Dependencies
+
+- Updated config to 0.4.0
+
+## 0.2.1 (2025-10-07)
+
+### 🧱 Updated Dependencies
+
+- Updated openai to 0.3.0
+
 ## 0.2.0 (2025-10-06)
 
 ### 🚀 Features